# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Visca Corporation API",
        "description": "Dokumentasi Open Api Visca Corporation",
        "contact": {
            "email": "bimasaktiputra95@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Local"
        }
    ],
    "paths": {
...
        "/v1/otentikasi/masuk": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "User Login",
                "description": "Login pengguna qollega (owner, pic, kasir, staff)",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Akun email aktif yang sudah di aktivasi",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Password akun",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Gagal login, akun belum diverifikasi / Akun anda belum terdaftar di sistem kami"
                    }
                }
            }
        },
        "/v1/otentikasi/daftar": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Pendaftaran Akun",
                "description": "Pendaftaran Akun Merchant",
                "operationId": "Pendaftaran Akun Merchant",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama Lengkap Pengguna",
                                        "type": "text"
                                    },
                                    "email": {
                                        "description": "Alamat Email yang akan di daftarkan",
                                        "type": "text"
                                    },
                                    "password": {
                                        "description": "Password Pengguna",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil mendaftarkan akun, silahkan login",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Pendaftaran gagal, periksa kembali data validasi anda",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}